Fix post request issue
One little missing return statement caused so much pain. The submit button
was causing the page to refresh too soon and stopping js that was running,
so the server was responding, just a tiny smidgen too slow before the page
refreshed. So the return false stops the page from refreshing and the
callback function runs, rerouting us to the landing page. About to cry.